Todos los casos vacíos

+++ remove_containers.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
 cd /code/gradio-docker-do
 [ -n "$(docker ps -a -q)" ] && docker stop $(docker ps -a -q)
-#docker stop $(docker ps -a -q)
-docker rm $(docker ps -a -q)
+[ -n "$(docker ps -a -q)" ] && docker rm $(docker ps -a -q)
